test/reftest/web-platform-tests/dom/events/Event-dispatch-propagation-stopped.html

<!DOCTYPE html>
<html>
<head>
<!--EDIT-->
<!--title> Calling stopPropagation() prior to dispatchEvent() </title-->
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #table {
    display: none;
  }
  .col {
    display: inline-block;
  }
  .cell {
    border: 1px solid black;
    padding: 0;
    margin: 1px;
  }
</style>
</head>
<body>
<div id=log></div>

<div id="table">
  <div id="table-body">
    <div class="col" id="parent">
      <p class="cell">Shady Grove</p>
      <p class="cell" id="target">Over the river, Charlie</p>
    </div>
    <div class="col">
      <p class="cell">Aeolian</p>
      <p class="cell">Dorian</p>
    </div>
  </div>
</div>

<script>
test(function() {
    var event = "foo";
    var target = document.getElementById("target");
    var parent = document.getElementById("parent");
    var tbody = document.getElementById("table-body");
    var table = document.getElementById("table");
    var body = document.body;
    var html = document.documentElement;
    var current_targets = [window, document, html, body, table, tbody, parent, target];
    var expected_targets = [];
    var actual_targets = [];
    var expected_phases = [];
    var actual_phases = [];

    var test_event = function(evt) {
        actual_targets.push(evt.currentTarget);
        actual_phases.push(evt.eventPhase);
    };

    for (var i = 0; i < current_targets.length; ++i) {
        current_targets[i].addEventListener(event, test_event, true);
        current_targets[i].addEventListener(event, test_event, false);
    }

    // EDIT
    // var evt = document.createEvent("Event");
    // evt.initEvent(event, true, true);
    var evt = new Event(event, {"bubbles":true, "cancelable":true});
    evt.stopPropagation();
    target.dispatchEvent(evt);

    assert_array_equals(actual_targets, expected_targets, "actual_targets");
    assert_array_equals(actual_phases, expected_phases, "actual_phases");
});
</script>
</body>
</html>
